o1vm: only legacy version should be using bn254
The pickles flavor would use Pasta curves.
